Tarea 4 de mayo

+++ ecommerceusb/src/main/java/co/edu/usbcali/ecommerceusb/repository/InventoryMovementRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InventoryMovementRepository extends JpaRepository<InventoryMovement, Long> {
+public interface InventoryMovementRepository extends JpaRepository<InventoryMovement, Integer> {
 }
